Ahora si hicimos algo decente. lectura de temo y hum

+++ GccApplication3/Include/Src/tempHumReg.c
@@ -8,18 +8,17 @@ extern volatile uint8_t Flag_SendData;
 extern char msg2[];
 
 void Task_TemHum(void) {
-	uint8_t temperature = 0;
-	uint8_t humidity = 0;
+	float temperature = 0;
+	float humidity = 0;
 
 	if (DHT11_Read(&temperature, &humidity) == 0) {
-		// while (1); nunca entra aca
 		char buffer[100];
 		uint8_t hour, minute, second, day, month, year;
 		DS3231_GetTime(&hour, &minute, &second);
 		DS3231_GetDate(&day, &month, &year);
 
 		snprintf(buffer, sizeof(buffer),
-		"TEMP: %d°C HUM: %d%% FECHA: %02d/%02d/%02d HORA: %02d:%02d:%02d\r\n",
+		"TEMP: %f°C HUM: %f%% FECHA: %02d/%02d/%02d HORA: %02d:%02d:%02d\r\n",
 		temperature, humidity, day, month, year, hour, minute, second);
 
 		if (Flag_SendData) {
